backend/requirements.txt
fastapi
uvicorn
sqlmodel
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0
python-jose[cryptography]
python-dotenv
pydantic-settings
psycopg2-binary
